finished subCategory on crud operation

+++ src/app/components/add-sub-category/add-sub-category.component.ts
@@ -7,6 +7,7 @@ import { FormsModule } from '@angular/forms';
 import { CategoryService } from '../../services/categoryService/category.service';
 import { Category } from '../../models/category/category';
 import { UpdatedSubCategoryDTO } from '../../models/DTOs/requestDTO/updatedSubCategoryDTO/updated-sub-category-dto';
+import { HandleResponse } from '../../handlingResponse/handle-response';
 
 @Component({
   selector: 'app-add-sub-category',
@@ -40,12 +41,19 @@ export class AddSubCategoryComponent {
       },
     });
   }
-  onSubmit(): void {
-    this.subCategoryService.addSubCategory(this.addedSubCategoryDTO).subscribe({
-      next: (data) => {
-        this.subCategories.push(data.entity);
-      },
-    });
+  async onSubmit() {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to add this sub category ?'
+    );
+    if (confirmed) {
+      this.subCategoryService
+        .addSubCategory(this.addedSubCategoryDTO)
+        .subscribe({
+          next: (data) => {
+            this.subCategories.push(data.entity);
+          },
+        });
+    }
   }
   getCategories(category: Category): Category[] {
     let categories: Category[] = [];
@@ -59,34 +67,44 @@ export class AddSubCategoryComponent {
     });
     return categories;
   }
-  deleteSubCategory(id: number) {
-    this.subCategoryService.deleteSubCategory(id).subscribe({
-      next: (data) => {
-        this.categories.map((category) => {
-          if (category.subCategories.length > 0) {
-            category.subCategories.forEach((subCategory) => {
-              if (subCategory.id == id) {
-                category.subCategories.splice(
-                  category.subCategories.indexOf(subCategory),
-                  1
-                );
-                return;
-              }
-            });
-          }
-        });
-      },
-    });
+  async deleteSubCategory(id: number) {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to delete this sub category ?'
+    );
+    if (confirmed) {
+      this.subCategoryService.deleteSubCategory(id).subscribe({
+        next: (data) => {
+          this.categories.map((category) => {
+            if (category.subCategories.length > 0) {
+              category.subCategories.forEach((subCategory) => {
+                if (subCategory.id == id) {
+                  category.subCategories.splice(
+                    category.subCategories.indexOf(subCategory),
+                    1
+                  );
+                  return;
+                }
+              });
+            }
+          });
+        },
+      });
+    }
   }
-  editSubCategory(subCategory: SubCategory) {
-    let UpdatedSubCategoryDTO: UpdatedSubCategoryDTO = {
-      Id: subCategory.id,
-      categoryId: subCategory.categoryId,
-      name: subCategory.name,
-    };
-    this.subCategoryService
-      .editSubCategory(UpdatedSubCategoryDTO)
-      .subscribe({});
+  async editSubCategory(subCategory: SubCategory) {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to edit this sub category ?'
+    );
+    if (confirmed) {
+      let UpdatedSubCategoryDTO: UpdatedSubCategoryDTO = {
+        Id: subCategory.id,
+        categoryId: subCategory.categoryId,
+        name: subCategory.name,
+      };
+      this.subCategoryService
+        .editSubCategory(UpdatedSubCategoryDTO)
+        .subscribe({});
+    }
   }
   getFilteredCategories(gender: string, collection: string): Category[] {
     let categories: Category[] = [];
@@ -100,20 +118,26 @@ export class AddSubCategoryComponent {
     });
     return categories;
   }
-  addSubCategory() {
-    this.subCategoryService.addSubCategory(this.addedSubCategoryDTO).subscribe({
-      next: (data) => {
-        this.categories.map((category) => {
-          if (category.id == this.addedSubCategoryDTO.categoryId) {
-            category.subCategories.push(data.entity);
-            return;
-          }
-        });
-        this.gender = '';
-        this.collection = '';
-        this.addedSubCategoryDTO.categoryId = 0;
-        this.addedSubCategoryDTO.name = '';
-      },
-    });
+  async addSubCategory() {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to add this sub category ?'
+    );
+    if (confirmed) {
+      this.subCategoryService.addSubCategory(this.addedSubCategoryDTO).subscribe({
+        next: (data) => {
+          this.categories.map((category) => {
+            if (category.id == this.addedSubCategoryDTO.categoryId) {
+              category.subCategories.push(data.entity);
+              return;
+            }
+          });
+          this.gender = '';
+          this.collection = '';
+          this.addedSubCategoryDTO.categoryId = 0;
+          this.addedSubCategoryDTO.name = '';
+        },
+      });
+  
+    }
   }
 }
