+++ src/app/components/edit-brand/edit-brand.component.ts
@@ -5,6 +5,7 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 import { AddedBrandDTO } from '../../models/DTOs/requestDTO/addedBrandDTO/added-brand-dto';
 import { UpdatedBrandDTO } from '../../models/DTOs/requestDTO/updatedBrandDTO/updated-brand-dto';
+import { HandleResponse } from '../../handlingResponse/handle-response';
 
 @Component({
   selector: 'app-edit-brand',
@@ -22,29 +23,44 @@ export class EditBrandComponent {
       this.brands = response.brands;
     });
   }
-  addBrand() {
-    let addedBrandDTO: AddedBrandDTO = {
-      name: this.newBrand,
-    };
-    this.brandService.addBrand(addedBrandDTO).subscribe({
-      next: (response) => {
-        this.brands.push(response.entity);
-        this.newBrand = '';
-      },
-    });
+  async addBrand() {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to add this brand ?'
+    );
+    if (confirmed) {
+      let addedBrandDTO: AddedBrandDTO = {
+        name: this.newBrand,
+      };
+      this.brandService.addBrand(addedBrandDTO).subscribe({
+        next: (response) => {
+          this.brands.push(response.entity);
+          this.newBrand = '';
+        },
+      });
+    }
   }
-  editBrand(brand: Brand) {
-    let updatedBrandDTO: UpdatedBrandDTO = {
-      name: brand.name,
-      Id: brand.id,
-    };
-    this.brandService.updateBrand(updatedBrandDTO).subscribe({});
+  async editBrand(brand: Brand) {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to edit this brand ?'
+    );
+    if (confirmed) {
+      let updatedBrandDTO: UpdatedBrandDTO = {
+        name: brand.name,
+        Id: brand.id,
+      };
+      this.brandService.updateBrand(updatedBrandDTO).subscribe({});
+    }
   }
-  deleteBrand(brandId: number) {
-    this.brandService.deleteBrand(brandId).subscribe({
-      next: (response) => {
-        this.brands = this.brands.filter((brand) => brand.id !== brandId);
-      },
-    });
+  async deleteBrand(brandId: number) {
+    const confirmed = await HandleResponse.operationConfirmed(
+      'Are you sure you want to delete this brand ?'
+    );
+    if (confirmed) {
+      this.brandService.deleteBrand(brandId).subscribe({
+        next: (response) => {
+          this.brands = this.brands.filter((brand) => brand.id !== brandId);
+        },
+      });
+    }
   }
 }
